Duzo zmian w room ale hatoes jeszcze nie dziala

+++ mobile-api/src/main/java/com/teamE/rooms/Room.java
@@ -37,7 +37,7 @@ public class Room {
     @Builder.Default
     private RoomConfiguration configuration = RoomConfiguration.getDefaultConfiguration();
 
-    private Integer mainImage;
+    private Long mainImage;
 }
